CNN/cifar_classify/requirements.txt
absl-py==0.3.0
appnope==0.1.0
astor==0.7.1
backcall==0.1.0
beautifulsoup4==4.6.3
bleach==2.1.3
bs4==0.0.1
certifi==2018.8.24
chardet==3.0.4
coremltools==2.0
cycler==0.10.0
decorator==4.3.0
entrypoints==0.2.3
gast==0.2.0
grpcio==1.14.1
h5py==2.8.0
html5lib==1.0.1
idna==2.7
ipykernel==4.8.2
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.0
jedi==0.12.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.3
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.6.0
numpy==1.14.5
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
prometheus-client==0.3.1
prompt-toolkit==1.0.15
protobuf==3.6.0
ptyprocess==0.6.0
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.3.1
requests==2.19.1
scikit-learn==0.19.2
scipy==1.1.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.10.0
sklearn==0.0
tensorboard==1.10.0
tensorflow_gpu==1.10.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
tfcoreml==0.2.0
tornado==5.1
traitlets==4.3.2
urllib3==1.23
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.0
